spotippos/properties/tests/__init__.py

from rest_framework.test import APIClient
from rest_framework.test import APIRequestFactory
from rest_framework.test import APITestCase

from properties.models import Property


class BaseApiTestCase(APITestCase):

    def setUp(self):
        self.factory = APIRequestFactory()
        self.client = APIClient()

    def tearDown(self):
        Property.objects.all().delete()
