spotippos/requirements.txt
coverage==4.0.3
Django==1.9
django-nose==1.4.2
djangorestframework==3.3.1
mock==1.3.0
nose==1.3.7
requests==2.8.1
